+++ packages/node-firebird-driver-native/src/lib/fb-util.ts
@@ -15,12 +15,12 @@ export function fixMetadata(status: fb.Status, metadata?: fb.MessageMetadata): f
 	try {
 		for (let i = metadata.getCountSync(status) - 1; i >= 0; --i) {
 			switch (metadata.getTypeSync(status, i)) {
-				// Transforms CHAR descriptors to VARCHAR.
+				// Transform CHAR descriptors to VARCHAR.
 				case sqlTypes.SQL_TEXT:
 					outBuilder.setTypeSync(status, i, sqlTypes.SQL_VARYING);
 					break;
 
-				// Transforms numeric descriptors to DOUBLE PRECISION.
+				// Transform numeric descriptors to DOUBLE PRECISION.
 				case sqlTypes.SQL_SHORT:
 				case sqlTypes.SQL_LONG:
 				case sqlTypes.SQL_INT64:
@@ -30,10 +30,12 @@ export function fixMetadata(status: fb.Status, metadata?: fb.MessageMetadata): f
 					outBuilder.setScaleSync(status, i, 0);
 					break;
 
-				// Transforms INT128 descriptors to VARCHAR.
+				// Transform INT128, DEC16 and DEC34 descriptors to VARCHAR.
 				case sqlTypes.SQL_INT128:
+				case sqlTypes.SQL_DEC16:
+				case sqlTypes.SQL_DEC34:
 					outBuilder.setTypeSync(status, i, sqlTypes.SQL_VARYING);
-					outBuilder.setLengthSync(status, i, 39);
+					outBuilder.setLengthSync(status, i, 45);
 					outBuilder.setCharSetSync(status, i, charSets.ascii);
 					break;
 			}